requirements.txt
streamlit
pysocks